Rework message format to have command type in header
Instead of putting the command type in the message args, use a single
byte in the header. The message variant is then simply the args and can
be an 'a{sv}' type. This saves a lot of bits since GVariant pads out the
command type quite a bit even when a byte type is used.