Add convenience methods for putobject
These methods handle receiving the actual putobject command as well as
writing or reading the subsequent object payload. On the receiving end,
the data handling is split out in a separate method to allow the caller
to check the message's type before receiving the payload.

The receiver writes the received objects into the tmp directory. Later
this probably needs to be handle like commit with full transaction
handling in a unique staging directory.