Send and receive all needed objects
Wire up the putobject handling into the pusher and the receiver. All
objects are pushed in the random order that ostree_traverse_commit()
gave them to us. After each putobject, the receiver sends back a
pass/fail status command. After all objects have been sent, the pusher
sends a done command to the receiver.

To handle receiving either a putobject or a done command, the receiver
first reads the message and then reads the pubobject data payload
separately if a done wasn't received.