Store the objects and update the refs to complete the push
After receiving all the objects, the receiver moves them from the
temporary directory to the object store. Finally, the refs are updated
to the commits specified in preceding update command.

This needs work to be ordered correctly so that the store remains valid
at all times. For instance, the commit object should be moved in last
after all it's descendents are in place. Otherwise, we'll need to fsck
all the time.