OSTreeReceiver: Use a session specific temp directory
This avoids cases where concurrent pushes contain files
which have the same checksum, which happens a lot with dirmeta
objects.

Specifically, what will happen is that two pushes will both
write the same dirmeta file, the first one will rename it
into the ostree repo, and the second one will fail because
it's source file has disappeared.

Also cleanup temporary caches when an error occurs; dont leave a
bunch of files lying around in a temp directory if the connection was
interrupted or aborted by the remote (or if a bug occurred).