Add new otpush package
This is a rewrite of how ostree-push and ostree-receive work. Rather
than a custom protocol that needs to stay in sync with how ostree
manages a repo, it reverses the process and initiates a pull from the
remote host. We've been using a version of this internally at Endless
for several years and have had almost no issues with it.

The code has been put into a python package primarily to facilitate
testing, but also to allow the core functionality to be reusable should
anyone want to write their own frontends. Console scripts will
eventually replace the existing protocol script.