Add tox configuration
This makes it easier to test the generated package and use different
python environments. Currently just a single python3 environment is
tested.

When scripts are run either directly through subprocess or via ssh, they
need to use programs from the tox virtualenv `bin` directory. A
`TOXBINDIR` environment variable is used to communicate it and `PATH` is
adjusted when it's present. However, when tox is in use `PYTHONPATH`
doesn't need to be adjusted since the `otpush` package will be installed
in an appropriate path.