Remove old protocol version and experimental sshfs script
This was a fun experiment but ultimately it would have needed a lot more
attention to work well. The new implementation uses more environmental
scaffolding, but ultimately it's simpler and requires less maintenance
as ostree itself is doing the heavy lifting.

This code is still available in the old-protocol branch should anyone
want to keep using it.