receive: Add configuration class
Add the `OTReceiveConfig` class to store `ostree-receive` configuration.
The config fields are a subset of those provided by the CLI and later
fields can be added that aren't exposed to the CLI. A `load` class
method will read values from a YAML formatted file and create an
`OTReceiveConfig` instance.

Since the `OTReceiveConfig` class uses `dataclasses`, the minimum python
version is raised to 3.7. The `PyYAML` package is required for reading
config files. Unfortunately, that means the ubuntu-18.04 tests need to
be dropped since only python 3.6 is available there. Probably the tests
should move to containers.

This is currently unused but will be combined with the CLI argument
parsing to configure `OTReceiveRepo`.