receive: Add support for running hook after receiving commits
In most cases, running either `ostree summary -u` or `flatpak
build-update-repo` will suffice to update the repo metadata after
receiving commits. In some cases this may not do all that the operator
needs and they need to run a custom routine.

The `update_hook` configuration parameter allows defining a program to
run after receiving commits. When the program is run, the
`OSTREE_RECEIVE_REPO` and `OSTREE_RECEIVE_REFS` environment variables
will be set as described in the example configuration.