github: Switch to docker containers
This is roughly the same as what was there but running in a container
instead of directly on the VM. A more detailed matrix configuration is
used to allow more images to be used from different distros.

One annoying issue encountered is that sshd refuses to start if
`/run/sshd` doesn't exist. When `openssh-server` is installed on the VM,
it starts `ssh.service`, which creates `/run/sshd`. That doesn't happen
in a container and the directory needs to be created. This is also not
likely to happen on Fedora where services are never started after
install.