receive: Factor out repo config from receiver config
Having the `OTReceiveRepo` class handle its own path resolution and all
the global configuration always felt wrong. This splits the repo
configuration from the receiver configuration with additional classes.

Now the repo class works entirely standalone on inputs that are relevant
to it. The receiver class is a small wrapper whose job is entirely to
translate from the global configuration to a specific repository
receiving commits.

Unfortunately, this duplicates the configuration fields and defaults,
but I couldn't think of a better way to handle it.