receive: Add support for per-repository configuration
In most cases the desired configuration for each repository is likely to
be the same. However, in some cases different may require different
settings, such as which key IDs to sign commits with.

Add a new `repos` configuration option to allow overriding the global
settings on a per-repository basis. The value is a map with repository
path keys and a map of override settings for values.