Reintroduce compatibility with old 0.x clients
Currently the command that gets launched on the receiver side is
installed as `ostree-receive`, as it was before the transition to the
new reverse-pull-over-http protocol.

This means that once the server is upgraded, older clients break badly
since the receiver side no longer understands their protocol.

To address this issue and retain compatibility with old clients:

1. install the new server side code with an unambiguous name,
   `ostree-receive-1`
2. make the new client do a lookup to see if `ostree-receive-1` is
   available before falling back to the ambiguous `ostree-receive` name
3. revive the old 0.x code and install it as `ostree-receive-0`
4. add a `ostree-receive` script to exec the right version depending on
   the invocation parameters (the `--repo` parameter is only used by the
   old code)

Signed-off-by: Emanuele Aina <emanuele.aina@collabora.com>