receive: Install and allow ostree-receive-1 alias
The client/server interface is dependent on the `ostree-receive` CLI. In
order to allow updating the CLI interface while supporting older clients
using the current interface, install `ostree-receive` with the major
version suffixed. I don't expect to change that interface, but this will
help support updating the client and server independently.

When pushing, prefer the `ostree-receive-<major>` command. This allows
the `ostree-receive-<major>` command matching the current major version
to be used if the server supports it. If not, the client will fall back
to the unversioned `ostree-receive`.