push: Allow skipping missing remote commands
In order to be flexible in how the receive command is deployed on the
server, allow the client to specify multiple commands when pushing. If
multiple commands are specified, each command is attempted until one is
found on the remote host. The default remains to use `ostree-receive` as
the single receive command.

The `push_refs` function is changed to use a `commands` parameter that
expects an iterable of command strings. The CLI interface is unchanged
but allows specifying multiple `--command` options in order of
preference.