receive: Add compatibility entry point for unversioned receiver
In order to handle both legacy and current clients executing the
unversioned `ostree-receive`, add a compatibility entry point for that
program which dispatches to the legacy entry point as needed. The
heuristic to see if the legacy receive is needed is whether the `--repo`
option is present. The repo path is a positional argument with the
current receive.

Fixes: #11